requirements.txt
asgiref==3.8.1
Django==5.0.7
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1
